Initial seeding of Searchlight UI
This repository is being seeded with a patch that has lived
on Horizon master for the Liberty and Mitaka cycles.

At the horizon mitaka mid-cycle, we decided to move this to a plugin
for the Mitaka release. It was not possible to do so until
several patch dependencies merged on horizon master. Those
have now merged, so we are able to put this into its own repo.