Add getProject policy check to project name filter
The project column is enabled for admins by default, but can be enabled
via searchlight_policy.json for all users.  When it is enabled, it will
cause a lookup for the project name to be done based on the project id
of the resource in the results. However, the keystone get_project
api itself has a policy on it and any requests for a project name
will fail if the current user doesn't have access rights to it.
By default, non-admin users will get a forbidden result which causes
horizon to log the current user out (horizon bug).

This patch adds a policy check for 'identity', 'identity:get_project'
before attempting to look up the project name.

Implements-blueprint horizon-initial-search-panel-plugin
Change-Id: I07123ea439fa0643fa73297a98fc70316414ab0e