Searchlight UI Query Error Handling
Currently, if there is an error with the query
for any reason, you get a blank row with no information.

This patch adds a warning message for query errors.

This occurs either due to malformed full text search
or due to bugs in the magic search to searchlight query
parsing. In either case, the user needs to know whether
or not they have zero results or if they are seeing an
error. This provides a catch all visualization to
guide users.

Screenshot examples:

Invalid Query:
http://pasteboard.co/2F74wVWZ.png

Turn off searchlight-api:
http://pasteboard.co/2GlfQIUO.png

Partially-Implements-blueprint horizon-initial-search-panel-plugin

Change-Id: Ib1ec73caa42fd4e329b548aed09aeefc7c10ff9d