Fix issues when magic search facet uses range queries
The facets returned from searchlight include their type.
The searchlight query generator is not using the correct
type of query at times. For example, a volume.size can do
range queries, but the searchlight UI query generator is
using a term query, resulting in a BadRequest from elasticsearch.

This fix changes it so that we treat all facets as
field limited query strings. This will now ensure
that the search syntax for full text search is
consistent with what you can put in for selecting
a single facet.

Example tests:

Select Type: Volume
Select Type: Size
 Inputs (try one at a time):
 >1
 >=1
 [1 TO 10]
 [1 TO *]
 >=1= (should result in a message that the query is invalid

Next, put a phrase into a volume description such as:

"Running this test = fun times"

Then do a bunch of tests on selecting the description field
and trying phrase and wildcard queries.

Try others.

In general, this seems to work much better than before with
consistency between the inputs on full text search and facet
limited search.

Change-Id: Ic126eae0ae7a422e5ac6d7be98a2bea57c13fc6f
Closes-Bug: 1564147