Remove unused hz-array-field-directive
The hz-array-field directive provided some additional syntax
highlighting taking nested table like objects and displaying
them as a table. The output format was interesting, but we
commented it out because it wasn't quite right. This patch
removes the commented out code for that purpose.
It can be added back at a later point if desired.

Change-Id: I3b7d20f147aaccf291aa20b4081aec57e1a09ae8
Closes-Bug: 1565947