Populate text search if saved
The UI doesn't currently populate the text search even when it is saved
and retrieved (and run).  This patch simply places the text in the input
where it belongs.

This is hard to test until we have SPA-routing, but it does work as shown
in GitHub: https://github.com/palecrow/searchlight-ui

Change-Id: Ife3a1930f598685419b3f4052db5c4130bfe5c64
Closes-Bug: 1573078