Code organization cleanup
Use a consistent "sl" prefix for directives and filters to avoid
potential collisions with Horizon or other UI frameworks that might
attempt to use this UI.

Flatten the directory structure and change the module names to match.

Consistently use the module namespace on injectables.

Currently the javascript files are organized according to a Horizon
dashboard pattern and aren't consistently namespaced.  There is little
reason to have the plugin structure follow the Horizon pattern for a
particular dashboard plugin, especially when considering searchlight-ui
may be placed in Project, Admin or outside of any existing dashboard, or
potentially used outside of Horizon entirely.

Implements: blueprint searchlight-ui-modularize
Change-Id: Ib2c906ff0d17f94eb2c668098716e12da13a0bde