Adding registry-based drawer (summary) views
This patch adds registry-based drawer views to the searchlight table.  This
means that for each resource type, the registry sees if there is a registered
summary view; if there is then it is shown, and if it is not present the
default view is presented.

The patch breaks out the views and also uses sl-drawer to encapsulate the
view, to be in alignment with the expectation of drawers (that they have an
'item' on their scope, etc.).

Change-Id: Id5833b00fe52fda0ce3dfcc4f9052fe10e7ec69f
Partially-Implements: blueprint sl-ui-use-hz-registry