Add Global Actions to search page
As demonstrated in the GitHub repo, we should have global actions showing up
in the search-bar area of the search page.  This patch allows this to happen.
Please note, you need some global actions to be registered in order for this
to show up; for example, in Horizon you can use:

https://review.openstack.org/#/c/334507/

Which should show Create Image at least in the list.

Change-Id: I27fceabc38932922b0961bde61c79f9f352e4474
Partially-Implements: blueprint sl-ui-use-hz-registry