Show items in-transition
Use the registry to detect items that are in a transition state
(like 'queued' images). Show a progress bar in the status column.

Also, for consistency, use that same progress bar in the actions
column for "dirty" items that the user has just acted upon.

Dirty items are polled individually at high frequency to give the
user a sense that their action has immediate effect.

In-transition items may appear as a result of this user's action,
or another user's action, so they are not individually polled.

A possible optimization is to individually poll in-transition items
that are a result of a user's action to dirty them...but this has
a variety of corner cases to consider and is not attempted as
part of this patch.

Partially-Implements: blueprint edit-live-data
Change-Id: If05789a8ce308537156e86085682943c1325a114