Display correct number of total results
Due to a dependency on smart table in horizon,
the number of results displayed is not accurate.

This patch now shows the total number of hits and
the total number that are displayed (limited by the search limit).

Pagination is still handled through smart table. This will
be fixed in a future patch.

Change-Id: Ia00de35a76e4259f77e32581871bf3614e381dcc
Closes-Bug: 1616754