Default handling of paths to objects
We can't tell if an object is nested or is an object.
Previously, we used nested as the default best guess.
Now we need to use object syntax.  The service really
needs to tell UI if a field is nested or not. This is
intended to deal with the ramifications of:

I35cb9b43d4281c2cabdb602bf64af096ca4b9955

Change-Id: I6102fa26450c255707008ba9388f29b2bd839abe
Closes-Bug: #1621289
Depends-On: I35cb9b43d4281c2cabdb602bf64af096ca4b9955