Save SL-UI settings separated by user
This patch uses the domain/username as part of the key for local storage.  NOTE:
this does not mean that the data is not available to the other users, however,
it is a good feature that allows different logged-in users to maintain different
settings.  This is also useful for an operator who uses different users to perform
different activities (monitoring, etc.).

Change-Id: Iab0ca354b72d6d70701e4f33cd7b0ff56ed9a94c
Closes-Bug: 1617369